feat: last edited timestamp in flows (#4226)
