+++ editor.planx.uk/src/pages/FlowEditor/utils.ts
@@ -36,3 +36,8 @@ export const formatLastEditMessage = (
 
 export const formatLastPublishMessage = (date: string, user: string): string =>
   `Last published ${formatLastEditDate(date)} by ${user}`;
+
+export const formatServiceLastUpdated = (date: string): string => {
+  const formattedDate = new Date(date).toLocaleDateString("en-gb");
+  return `Service last updated ${formattedDate}`;
+};
